+++ src/3rdparty/glfw/include/GLFW/glfw3native.h
@@ -1,0 +1,3 @@
+// Compatibility forwarding header. See glfw3.h in this directory.
+#pragma once
+#include "../glfw3native.h"